Breaks out build_recovery_store
